feat : 2021.03.22 Exception 처리

+++ src/main/java/com/example/escape/service/ApiService.java
@@ -15,24 +15,67 @@ public class ApiService {
     ApiDao apiDao;
 
 
+    /**
+     * 모든 list 조회
+     * @return
+     */
     public List<Product> getAllProducts() {
         return apiDao.getAllProducts();
     }
 
-    public Product getProductByPdId(String pdId) {
+    /**
+     * id 로  데이터 조회 
+     * @param pdId
+     * @return
+     * @throws Exception
+     */
+    public Product getProductByPdId(String pdId) throws Exception {
         return apiDao.getProductByPdId(pdId);
     }
 
-    public Product registerProduct(Product product) {
+    /**
+     * product 상품 등록
+     * @param product
+     * @return
+     * @throws Exception
+     */
+    public Product registerProduct(Product product) throws Exception {
+
+        if((isDuplicatedPdId(product.getPdId()))){
+            throw new RuntimeException("the Pdid already exixts");
+        }
+
         return apiDao.insertProduct(product);
     }
 
-    public void modifyProduct(String pdId, Product product) {
-        apiDao.updateProduct(pdId, product);
+    /**
+     *  product 데이터 수정
+     * @param pdId
+     * @param product
+     * @return
+     * @throws Exception
+     */
+    public Product modifyProduct(String pdId, Product product) throws Exception {
+        Product product1 = getProductByPdId(pdId);
+        if(!(isDuplicatedPdId(pdId))){
+            throw new RuntimeException("No Search Data by PdId");
+        }
+
+        return product1;
+
     }
 
-    public void removeProduct(String pdId) {
-        apiDao.deleteProduct(pdId);
+    /**
+     * pdid 로 해당 데이터 삭제
+     * @param pdId
+     * @throws Exception
+     */
+    public void removeProduct(String pdId)  throws Exception {
+        if(!(isDuplicatedPdId(pdId))){
+            throw new RuntimeException("No Search Data by PdId");
+        }
+         apiDao.deleteProduct(pdId);
+
     }
 
     public Map<String, String> getTest (){
@@ -40,5 +83,19 @@ public class ApiService {
         res.put("test", "hi");
 
         return res;
+    }
+
+
+    /**
+     * 변경하려는 pdId 이미 있는지 확인
+     *
+     * @param pdId
+     * @return boolean
+     */
+    private boolean isDuplicatedPdId(String pdId) throws Exception {
+        if (apiDao.getProductByPdId(pdId) != null) {
+            return true;
+        }
+        return false;
     }
 }
